Move git_info to top level of package module
Currently git_info resides in katsdpscripts.RTS and importing it forces
you to import the rest of the RTS module as well. This in turn exposes you
(via rfilib) to some weird and cranky dependencies such as pathos and skimage,
a far cry from basic package versioning.

Since git_info is so basic, rather move it one level higher to the clean
top-level package.
