+++ katsdpscripts/RTS/__init__.py
@@ -1,6 +1,5 @@
 from weatherlib import weather_report
 from rfilib import generate_flag_table, generate_rfi_report
-from git_info import git_info
 import diodelib                         # For QT 2_2
 import spectral_baseline                # For QT 2.10,3.8
 import strong_sources                   # For QT 2.8
