+++ katsdpscripts/__init__.py
@@ -1,3 +1,5 @@
+from .git_info import git_info
+
 # BEGIN VERSION CHECK
 # Get package version when locally imported from repo or via -e develop install
 try:
