+++ katsdpscripts/git_info.py
